+++ .eslintrc
@@ -15,6 +15,8 @@
     "react/destructuring-assignment": "ignoreClassFields",
     "react/forbid-foreign-prop-types": "off",
     "react/jsx-filename-extension": "off",
-    "prettier/prettier": ["error"]
+    "prettier/prettier": ["error"],
+    "import/prefer-default-export": "off",
+    "react/require-default-props": "off"
   }
 }
