+++ package.json
@@ -42,7 +42,7 @@
     "express": "^4.16.4",
     "highlight-words-core": "^1.2.2",
     "lodash.debounce": "4.0.8",
-    "prop-types": "^15.6.2",
+    "prop-types": "^15.7.2",
     "rc-input-number": "^4.5.1",
     "react-highlight-words": "^0.16.0",
     "react-switch": "4.1.0",
@@ -64,6 +64,8 @@
     "@semantic-release/changelog": "^3.0.1",
     "@semantic-release/git": "^7.0.5",
     "@semantic-release/npm": "^5.1.1",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.4.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
