+++ src/components/MenuItem/README.md
@@ -1,0 +1,2 @@
+# TODO: Documentation and purpose
+REMIND ME TO DO SOMETHING HERE :D
