+++ src/utils/propTypes/deprecatedPropType.js
@@ -1,0 +1,7 @@
+import { shouldNotBeDefinedPropType } from './base/shouldNotBeDefinedPropType';
+
+const deprecatedPropTypeFactory = shouldNotBeDefinedPropType('deprecated');
+
+export function deprecatedPropType(...args) {
+  return deprecatedPropTypeFactory(...args);
+}
