+++ src/utils/propTypes/index.js
@@ -1,0 +1,2 @@
+export * as deprecatedPropType from './deprecatedPropType';
+export * as unimplementedPropType from './unimplementedPropType';
