+++ src/utils/propTypes/unimplementedPropType.js
@@ -1,0 +1,9 @@
+import { shouldNotBeDefinedPropType } from './base/shouldNotBeDefinedPropType';
+
+const unimplementedPropTypeFactory = shouldNotBeDefinedPropType(
+  'not implemented yet'
+);
+
+export function unimplementedPropType(...args) {
+  return unimplementedPropTypeFactory(...args);
+}
